Move selected event to new file
