+++ core/block_svg.js
@@ -20,8 +20,8 @@ goog.require('Blockly.constants');
 goog.require('Blockly.ContextMenu');
 goog.require('Blockly.ContextMenuRegistry');
 goog.require('Blockly.Events');
-goog.require('Blockly.Events.OldUi');
 goog.require('Blockly.Events.BlockMove');
+goog.require('Blockly.Events.Selected');
 goog.require('Blockly.Msg');
 goog.require('Blockly.navigation');
 goog.require('Blockly.RenderedConnection');
